requirements.txt
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.2.1
colorama==0.4.6
cryptography==46.0.6
Flask==3.1.1
flask-cors==6.0.1
google-api-core==2.30.0
google-api-python-client==2.193.0
google-auth==2.49.1
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.3.0
googleapis-common-protos==1.73.1
httplib2==0.31.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
numpy==2.4.3
oauthlib==3.3.1
pandas==3.0.1
proto-plus==1.27.2
protobuf==6.33.6
py-cpuinfo==9.0.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.33.0
requests-oauthlib==2.0.0
six==1.17.0
sympy==1.14.0
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
